+++ changelog.md
@@ -121,6 +121,7 @@
 - URL sanitizer for last visited page tracking.
 - For security reasons, do not delete administrators via the customer grid. Double ask if administrators are deleted via the edit page.
 - Added filters to shipping by total and shipping by weight grids.
+- #862 Changed resources of X (Twitter) to X. 
 
 ### Bugfixes